requirements.txt
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
dotenv==0.9.9
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
playwright==1.52.0
playwright-stealth==1.0.6
pluggy==1.5.0
pyee==13.0.0
PySocks==1.7.1
pytest==8.3.5
pytest-base-url==2.1.0
pytest-playwright==0.7.0
python-dotenv==1.1.0
python-slugify==8.0.4
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.3
selenium==4.31.0
setuptools==80.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
text-unidecode==1.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
